requirements.txt
openslide-python
girder-client
tqdm